Partially working hints again.
Need to fix crashes on hint generation. May need to start from scratch.

+++ source/include/item_location.hpp
@@ -156,7 +156,7 @@ public:
       return scene;
     }
 
-    u8 GetTextID() const {
+    u16 GetTextID() const {
       return textID;
     }
